// pages/contentTypes.d.ts
export interface ApiBlogBlog extends Schema.CollectionType {
  collectionName: 'blogs';
  info: {
    singularName: 'blog';
    pluralName: 'blogs';
    displayName: 'Blog';
    description: '';
  };
  options: {
    draftAndPublish: true;
  };
  attributes: {
    Title: Attribute.String & Attribute.Required;
    Slug: Attribute.UID<'api::blog.blog', 'Title'> & Attribute.Required;
    PublishedDate: Attribute.Date & Attribute.Required;
    Content: Attribute.RichText & Attribute.Required;
    Image: Attribute.Media;
    Author: Attribute.Enumeration<
      [
        'Indra Kurniawan',
        'Izzul Qudsi',
        'John Papilon Steven Guntoro',
        'Mordekhai',
        'Muhamad Reza Noviardi',
        'Praviandy Gustandika',
        'Rahmantara Trichandi'
      ]
    > &
      Attribute.Required;
    ReadMoreURL: Attribute.String;
    seo: Attribute.Component<'shared.seo'>;
    createdAt: Attribute.DateTime;
    updatedAt: Attribute.DateTime;
    publishedAt: Attribute.DateTime;
    createdBy: Attribute.Relation<'api::blog.blog', 'oneToOne', 'admin::user'> &
      Attribute.Private;
    updatedBy: Attribute.Relation<'api::blog.blog', 'oneToOne', 'admin::user'> &
      Attribute.Private;
  };
}